+++ default.nix
@@ -6,12 +6,13 @@
 , template-haskell , these, time, transformers
 , transformers-compat, unbounded-delays, prim-uniq
 , data-default, filepath, directory, filemanip, ghcjs-base
-, monoidal-containers, witherable
+, monoidal-containers, witherable, profunctors
+, semialign ? null, splitThese ? (semialign != null)
 , useTemplateHaskell ? true
 }:
 mkDerivation {
   pname = "reflex";
-  version = "0.6.2.0";
+  version = "0.6.2.4";
   src = builtins.filterSource (path: type: !(builtins.elem (baseNameOf path) [ ".git" "dist" ])) ./.;
   libraryHaskellDepends = [
     base bifunctors containers dependent-map dependent-sum
@@ -22,17 +23,20 @@ mkDerivation {
     base bifunctors containers deepseq dependent-map dependent-sum
     mtl ref-tf split transformers data-default
     random time unbounded-delays monoidal-containers witherable
+    profunctors
   ] ++ (if ghc.isGhcjs or false then [
     ghcjs-base
   ] else []) ++ (if !useTemplateHaskell then [] else [
     haskell-src-exts haskell-src-meta
-  ]);
+  ]) ++ (if splitThese then [
+    semialign
+  ] else []);
   testHaskellDepends = if ghc.isGhcjs or false then [] else [
     hlint filepath directory filemanip
   ];
-  configureFlags = if useTemplateHaskell then [] else [
-    "-f-use-template-haskell"
-  ];
+  configureFlags =
+    stdenv.lib.optional (!useTemplateHaskell) [ "-f-use-template-haskell" ] ++
+    stdenv.lib.optional (!splitThese) [ "-f-split-these" ];
   homepage = "https://github.com/reflex-frp/reflex";
   description = "Higher-order Functional Reactive Programming";
   license = stdenv.lib.licenses.bsd3;